pyproject.toml
[project]
name = "git-change-detection"
version = "0.1.0"
description = "Detect which Ansible playbooks need to run based on git changes and metadata files"
readme = "README.md"
requires-python = ">=3.13"
license = {text = "MIT"}
dependencies = [
  "click>=8.1",
  "jinja2>=3.1.6",
  "pyyaml>=6.0",
  "gitpython>=3.1",
]

[project.scripts]
git-change-detection = "git_change_detection.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/git_change_detection"]
